requirements.txt
python-telegram-bot
container_app_conf
pydantic
emoji
